+++ example/custom/main.go
@@ -34,26 +34,13 @@ import (
 )
 
 func main() {
-	h := server.Default(
-		server.WithHostPorts("127.0.0.1:8080"),
-	)
+	h := server.New(server.WithHostPorts(":8080"))
 	h.Use(secure.New(
 		secure.WithAllowedHosts([]string{"example.com", "ssl.example.com"}),
-		secure.WithSSLRedirect(true),
 		secure.WithSSLHost("ssl.example.com"),
-		secure.WithSTSSecond(315360000),
-		secure.WithSTSIncludeSubdomains(true),
-		secure.WithFrameDeny(true),
-		secure.WithContentTypeNosniff(true),
-		secure.WithBrowserXssFilter(true),
-		secure.WithContentSecurityPolicy("default-src 'self'"),
-		secure.WithIENoOpen(true),
-		secure.WithReferrerPolicy("strict-origin-when-cross-origin"),
-		secure.WithSSLProxyHeaders(map[string]string{"X-Forwarded-Proto": "https"}),
 	))
-
-	h.GET("/ping", func(c context.Context, ctx *app.RequestContext) {
-		ctx.String(200, "pong")
+	h.GET("/ping", func(ctx context.Context, c *app.RequestContext) {
+		c.String(200, "pong")
 	})
 	h.Spin()
 }
